Bruteforce several steps
This version secured 4th place on Highest scores leaderboard.
Now I need to improve performance and parallel computations

+++ common/src/point.rs
@@ -7,13 +7,13 @@ pub struct Point {
 impl Point {
     pub const ZERO: Self = Point { x: 0, y: 0 };
 
-    pub fn len2(&self) -> f64 {
-        let x = self.x as f64;
-        let y = self.y as f64;
+    pub fn len2(&self) -> i64 {
+        let x = self.x as i64;
+        let y = self.y as i64;
         x * x + y * y
     }
     pub fn len(&self) -> f64 {
-        self.len2().sqrt()
+        (self.len2() as f64).sqrt()
     }
 
     pub fn scale(&self, target_len: f64) -> Self {
